src/collectors/news_collector.py
"""
News collection functionality
"""
from typing import List, Dict, Optional
from pathlib import Path
import json
import logging

class NewsDataset:
    """Handles news data collection and processing"""
    
    def __init__(self, api_key: str):
        self.api_key = api_key
        self.logger = logging.getLogger(__name__)

    def fetch_articles(self, force: bool = False) -> List[Dict]:
        """
        Fetch articles from the news API
        
        Args:
            force (bool): Force fetch even if cached
            
        Returns:
            List[Dict]: List of article data
        """
        self.logger.info("Fetch articles placeholder")
        return []

    def process_articles(self, articles: List[Dict]) -> List[Dict]:
        """
        Process fetched articles
        
        Args:
            articles (List[Dict]): Raw article data
            
        Returns:
            List[Dict]: Processed article data
        """
        self.logger.info("Process articles placeholder")
        return []

    def save_articles(self, articles: List[Dict]) -> Path:
        """
        Save processed articles
        
        Args:
            articles (List[Dict]): Processed article data
            
        Returns:
            Path: Path to saved data
        """
        self.logger.info("Save articles placeholder")
        return Path("placeholder_path") 